Splits each method in its own file
As it is the library has a hard dependency on certain node-only classes. This change allows for the separate require of each method, sidestepping some of these dependencies. For instance:

require('qr-image/lib/svg-object')
