+++ lib/image.js
@@ -1,0 +1,49 @@
+"use strict";
+
+var Readable = require('stream').Readable;
+
+var QR = require('./qr-base').QR;
+var png = require('./png');
+var vector = require('./vector');
+var get_options = require('./options');
+
+var fn_noop = function() {};
+
+function qr_image(text, options) {
+    options = get_options(options);
+
+    var matrix = QR(text, options.ec_level, options.parse_url);
+    var stream = new Readable();
+    stream._read = fn_noop;
+
+    switch (options.type) {
+    case 'svg':
+    case 'pdf':
+    case 'eps':
+        process.nextTick(function() {
+            vector[options.type](matrix, stream, options.margin, options.size);
+        });
+        break;
+    case 'svgpath':
+        // deprecated, use svg_object method
+        process.nextTick(function() {
+            var obj = vector.svg_object(matrix, options.margin, options.size);
+            stream.push(obj.path);
+            stream.push(null);
+        });
+        break;
+    case 'png':
+    default:
+        process.nextTick(function() {
+            var bitmap = png.bitmap(matrix, options.size, options.margin);
+            if (options.customize) {
+                options.customize(bitmap);
+            }
+            png.png(bitmap, stream);
+        });
+    }
+
+    return stream;
+}
+
+module.exports = qr_image;
